fix(reborn): address review feedback on loop refactor (#3904)
