+++ crates/ironclaw_agent_loop/src/executor/capabilities.rs
@@ -1,102 +1,8 @@
 use std::collections::HashSet;
 
-use super::input::*;
 use super::*;
 
 impl CanonicalAgentLoopExecutor {
-    pub(super) async fn stream_model_with_recovery(
-        &self,
-        planner: &dyn AgentLoopPlannerInternal,
-        host: &(dyn AgentLoopDriverHost + Send + Sync),
-        mut state: LoopExecutionState,
-        request: LoopModelRequest,
-    ) -> Result<ModelStep, AgentLoopExecutorError> {
-        let mut recorded_failure = false;
-        for _ in 0..MAX_MODEL_RETRIES {
-            match host.stream_model(request.clone()).await {
-                Ok(response) => {
-                    state.recovery_state = state.recovery_state.cleared_attempts();
-                    return Ok(ModelStep::Response(Box::new(state), response));
-                }
-                Err(error) => {
-                    if error.kind == AgentLoopHostErrorKind::Cancelled {
-                        return Err(AgentLoopExecutorError::Cancelled);
-                    }
-                    let Some(class) = model_error_class(&error) else {
-                        return Err(AgentLoopExecutorError::HostUnavailable {
-                            stage: HostStage::Model,
-                        });
-                    };
-                    if !recorded_failure {
-                        state.recent_failure_kinds.push(LoopFailureKind::ModelError);
-                        recorded_failure = true;
-                    }
-                    let summary = ModelErrorSummary {
-                        class,
-                        safe_summary: sanitized_strategy_summary(error.safe_summary)?,
-                        diagnostic_ref: error.diagnostic_ref,
-                    };
-                    match planner.recovery().on_model_error(&state, &summary).await {
-                        RecoveryOutcome::Retry {
-                            recovery, alter, ..
-                        } => {
-                            state.recovery_state = recovery;
-                            match self.checkpoint_and_exit_if_cancelled(host, state).await? {
-                                CancelCheck::Continue(next) => state = *next,
-                                CancelCheck::Exit(exit) => return Ok(ModelStep::Exit(exit)),
-                            }
-                            honor_retry_alteration(alter.as_ref())?;
-                            self.emit_progress(
-                                host,
-                                LoopProgressEvent::driver_note(
-                                    LoopDriverNoteKind::Retrying,
-                                    "retrying model request",
-                                )
-                                .map_err(|_| {
-                                    AgentLoopExecutorError::PlannerContract {
-                                        detail: "retry progress summary was invalid",
-                                    }
-                                })?,
-                            )
-                            .await;
-                        }
-                        RecoveryOutcome::SkipResult { .. } => {
-                            return Err(AgentLoopExecutorError::PlannerContract {
-                                detail: "SkipResult on model error",
-                            });
-                        }
-                        RecoveryOutcome::Abort {
-                            recovery,
-                            failure_kind,
-                        } => {
-                            state.recovery_state = recovery;
-                            match self.checkpoint_and_exit_if_cancelled(host, state).await? {
-                                CancelCheck::Continue(next) => state = *next,
-                                CancelCheck::Exit(exit) => return Ok(ModelStep::Exit(exit)),
-                            }
-                            let checked =
-                                self.checkpoint(host, state, CheckpointKind::Final).await?;
-                            return Ok(ModelStep::Exit(failed_exit(
-                                host,
-                                checked.state,
-                                failure_kind,
-                                Some(checked.checkpoint_id),
-                            )?));
-                        }
-                    }
-                }
-            }
-        }
-
-        let checked = self.checkpoint(host, state, CheckpointKind::Final).await?;
-        Ok(ModelStep::Exit(failed_exit(
-            host,
-            checked.state,
-            LoopFailureKind::DriverBug,
-            Some(checked.checkpoint_id),
-        )?))
-    }
-
     pub(super) async fn execute_capability_batch(
         &self,
         planner: &dyn AgentLoopPlannerInternal,
@@ -188,7 +94,8 @@ impl CanonicalAgentLoopExecutor {
             .await
             .map_err(capability_host_error)?;
 
-        if batch.outcomes.len() > visible_calls.len()
+        if batch.outcomes.is_empty()
+            || batch.outcomes.len() > visible_calls.len()
             || (!batch.stopped_on_suspension && batch.outcomes.len() != visible_calls.len())
         {
             return Err(AgentLoopExecutorError::PlannerContract {
@@ -210,7 +117,21 @@ impl CanonicalAgentLoopExecutor {
         )
         .await;
 
-        for (call, outcome) in visible_calls.into_iter().zip(batch.outcomes) {
+        let outcomes = batch.outcomes;
+        if !batch.stopped_on_suspension {
+            for (call, outcome) in visible_calls.iter().zip(&outcomes) {
+                if let CapabilityOutcome::Completed(result) = outcome {
+                    push_call_signature_once(&mut state, &mut signatures, call)?;
+                    append_capability_result_ref(host, call, result).await?;
+                    push_completed_result(&mut state, result.clone());
+                }
+            }
+        }
+
+        for (call, outcome) in visible_calls.into_iter().zip(outcomes) {
+            if !batch.stopped_on_suspension && matches!(outcome, CapabilityOutcome::Completed(_)) {
+                continue;
+            }
             push_call_signature_once(&mut state, &mut signatures, &call)?;
             match self
                 .handle_capability_outcome(planner, host, state, call, outcome)
@@ -330,15 +251,6 @@ impl CanonicalAgentLoopExecutor {
                 RecoveryOutcome::Retry {
                     recovery, alter, ..
                 } => {
-                    if matches!(summary.class, CapabilityErrorClass::PolicyDenied) {
-                        state.recovery_state = recovery;
-                        append_capability_error_ref(host, &mut state, &call, &summary).await?;
-                        match self.checkpoint_and_exit_if_cancelled(host, state).await? {
-                            CancelCheck::Continue(next) => state = *next,
-                            CancelCheck::Exit(exit) => return Ok(BatchStep::Exit(exit)),
-                        }
-                        return Ok(BatchStep::Continue(Box::new(state)));
-                    }
                     state.recovery_state = recovery;
                     match self.checkpoint_and_exit_if_cancelled(host, state).await? {
                         CancelCheck::Continue(next) => state = *next,
@@ -457,86 +369,6 @@ impl CanonicalAgentLoopExecutor {
             Some(checked.checkpoint_id),
         )?))
     }
-
-    pub(super) async fn handle_gate(
-        &self,
-        planner: &dyn AgentLoopPlannerInternal,
-        host: &(dyn AgentLoopDriverHost + Send + Sync),
-        mut state: LoopExecutionState,
-        call: CapabilityCallCandidate,
-        kind: GateKind,
-        gate_ref: ironclaw_turns::LoopGateRef,
-    ) -> Result<BatchStep, AgentLoopExecutorError> {
-        let summary = crate::strategies::GateSummary {
-            kind,
-            gate_ref: gate_ref.clone(),
-        };
-        match planner.gate().handle(&state, &summary).await {
-            GateOutcome::Block { gate } => {
-                state.gate_state = gate;
-                state.last_gate = Some(gate_ref.clone());
-                match self.checkpoint_and_exit_if_cancelled(host, state).await? {
-                    CancelCheck::Continue(next) => state = *next,
-                    CancelCheck::Exit(exit) => return Ok(BatchStep::Exit(exit)),
-                }
-                self.emit_progress(
-                    host,
-                    LoopProgressEvent::GateBlocked {
-                        iteration: state.iteration,
-                        gate_kind: loop_gate_kind(kind),
-                    },
-                )
-                .await;
-                let checked = self
-                    .checkpoint(host, state, CheckpointKind::BeforeBlock)
-                    .await?;
-                Ok(BatchStep::Exit(LoopExit::Blocked(LoopBlocked {
-                    kind: blocked_kind(kind),
-                    gate_ref,
-                    checkpoint_id: checked.checkpoint_id,
-                    state_ref: checked.state_ref,
-                    exit_id: exit_id(host, "blocked")?,
-                })))
-            }
-            GateOutcome::SkipAndContinue { gate } => {
-                state.gate_state = gate;
-                append_capability_safe_summary_ref(
-                    host,
-                    &mut state,
-                    &call,
-                    gate_tool_result_summary(kind, "skipped"),
-                )
-                .await?;
-                match self.checkpoint_and_exit_if_cancelled(host, state).await? {
-                    CancelCheck::Continue(next) => state = *next,
-                    CancelCheck::Exit(exit) => return Ok(BatchStep::Exit(exit)),
-                }
-                Ok(BatchStep::Continue(Box::new(state)))
-            }
-            GateOutcome::Abort { gate, failure_kind } => {
-                state.gate_state = gate;
-                append_capability_safe_summary_ref(
-                    host,
-                    &mut state,
-                    &call,
-                    gate_tool_result_summary(kind, "aborted"),
-                )
-                .await?;
-                match self.checkpoint_and_exit_if_cancelled(host, state).await? {
-                    CancelCheck::Continue(next) => state = *next,
-                    CancelCheck::Exit(exit) => return Ok(BatchStep::Exit(exit)),
-                }
-                let checked = self.checkpoint(host, state, CheckpointKind::Final).await?;
-                Ok(BatchStep::Exit(failed_exit(
-                    host,
-                    checked.state,
-                    failure_kind,
-                    Some(checked.checkpoint_id),
-                )?))
-            }
-        }
-    }
-
     pub(super) async fn fail_unsupported_process_wait(
         &self,
         host: &(dyn AgentLoopDriverHost + Send + Sync),
@@ -578,219 +410,5 @@ impl CanonicalAgentLoopExecutor {
             checked.state,
             Some(checked.checkpoint_id),
         )?))
-    }
-
-    pub(super) async fn exit_for_stop(
-        &self,
-        host: &(dyn AgentLoopDriverHost + Send + Sync),
-        state: LoopExecutionState,
-        kind: StopKind,
-    ) -> Result<LoopExit, AgentLoopExecutorError> {
-        match kind {
-            StopKind::GracefulStop => {
-                let checked = self.checkpoint(host, state, CheckpointKind::Final).await?;
-                completed_exit(host, checked.state, Some(checked.checkpoint_id))
-            }
-            StopKind::NoProgressDetected => {
-                let checked = self.checkpoint(host, state, CheckpointKind::Final).await?;
-                failed_exit(
-                    host,
-                    checked.state,
-                    LoopFailureKind::NoProgressDetected,
-                    Some(checked.checkpoint_id),
-                )
-            }
-            StopKind::Aborted(failure_kind) => {
-                let checked = self.checkpoint(host, state, CheckpointKind::Final).await?;
-                failed_exit(
-                    host,
-                    checked.state,
-                    failure_kind,
-                    Some(checked.checkpoint_id),
-                )
-            }
-        }
-    }
-
-    pub(super) async fn checkpoint(
-        &self,
-        host: &(dyn AgentLoopDriverHost + Send + Sync),
-        mut state: LoopExecutionState,
-        kind: CheckpointKind,
-    ) -> Result<CheckpointWrite, AgentLoopExecutorError> {
-        state.last_checkpoint = Some(crate::state::CheckpointMarker {
-            kind,
-            iteration_at_checkpoint: state.iteration,
-        });
-        let payload = serde_json::to_vec(&state)
-            .map_err(|_| AgentLoopExecutorError::CheckpointFailed { stage: kind })?;
-        let host_kind = checkpoint_kind_to_host(kind);
-        let state_ref = host
-            .stage_checkpoint_payload(StageCheckpointPayloadRequest {
-                kind: host_kind,
-                schema_id: crate::state::CHECKPOINT_SCHEMA_ID.to_string(),
-                payload,
-            })
-            .await
-            .map_err(|_| AgentLoopExecutorError::CheckpointFailed { stage: kind })?;
-        let checkpoint_id = host
-            .checkpoint(LoopCheckpointRequest {
-                kind: host_kind,
-                state_ref: state_ref.clone(),
-            })
-            .await
-            .map_err(|_| AgentLoopExecutorError::CheckpointFailed { stage: kind })?;
-        self.emit_progress(
-            host,
-            LoopProgressEvent::CheckpointWritten {
-                iteration: state.iteration,
-                kind: host_kind,
-            },
-        )
-        .await;
-        Ok(CheckpointWrite {
-            state,
-            checkpoint_id,
-            state_ref,
-        })
-    }
-
-    pub(super) async fn emit_progress(
-        &self,
-        host: &(dyn AgentLoopDriverHost + Send + Sync),
-        event: LoopProgressEvent,
-    ) {
-        let _ = host.emit_loop_progress(event).await;
-    }
-
-    // Cancellation is checked cooperatively at N boundary points between external calls.
-    // A macro refactor was considered but deferred; the explicit sites are self-documenting.
-    pub(super) async fn checkpoint_and_exit_if_cancelled(
-        &self,
-        host: &(dyn AgentLoopDriverHost + Send + Sync),
-        state: LoopExecutionState,
-    ) -> Result<CancelCheck, AgentLoopExecutorError> {
-        let Some(signal) = host.observe_cancellation() else {
-            return Ok(CancelCheck::Continue(Box::new(state)));
-        };
-
-        let fallback_state = state.clone();
-        match self.checkpoint(host, state, CheckpointKind::Final).await {
-            Ok(checked) => Ok(CancelCheck::Exit(cancelled_exit_with_reason(
-                host,
-                checked.state,
-                cancelled_reason_from_signal(&signal),
-                Some(checked.checkpoint_id),
-            )?)),
-            // Permissive profile: only checkpoint-write failures are absorbed
-            // into a checkpoint-free `Cancelled` exit. Other variants (e.g.
-            // `HostUnavailable`) must propagate so the runner can apply its
-            // recovery policy.
-            Err(AgentLoopExecutorError::CheckpointFailed { .. })
-                if !host
-                    .run_context()
-                    .resolved_run_profile
-                    .checkpoint_policy
-                    .require_final_checkpoint =>
-            {
-                Ok(CancelCheck::Exit(cancelled_exit_with_reason(
-                    host,
-                    fallback_state,
-                    cancelled_reason_from_signal(&signal),
-                    None,
-                )?))
-            }
-            Err(error) => Err(error),
-        }
-    }
-
-    pub(super) async fn checkpoint_and_exit_if_cancelled_after_pending_input_ack(
-        &self,
-        host: &(dyn AgentLoopDriverHost + Send + Sync),
-        state: LoopExecutionState,
-        pending_input_ack: &mut PendingInputAck,
-    ) -> Result<CancelCheck, AgentLoopExecutorError> {
-        let Some(signal) = host.observe_cancellation() else {
-            return Ok(CancelCheck::Continue(Box::new(state)));
-        };
-
-        let fallback_state = state.clone();
-        match self.checkpoint(host, state, CheckpointKind::Final).await {
-            Ok(checked) => {
-                pending_input_ack.ack(host).await?;
-                Ok(CancelCheck::Exit(cancelled_exit_with_reason(
-                    host,
-                    checked.state,
-                    cancelled_reason_from_signal(&signal),
-                    Some(checked.checkpoint_id),
-                )?))
-            }
-            // Permissive profile: absorb only checkpoint-write failures. The
-            // pending ack is intentionally NOT flushed here — no durable
-            // checkpoint was written, so advancing the input cursor would
-            // commit progress that the runner has no record of.
-            Err(AgentLoopExecutorError::CheckpointFailed { .. })
-                if !host
-                    .run_context()
-                    .resolved_run_profile
-                    .checkpoint_policy
-                    .require_final_checkpoint =>
-            {
-                Ok(CancelCheck::Exit(cancelled_exit_with_reason(
-                    host,
-                    fallback_state,
-                    cancelled_reason_from_signal(&signal),
-                    None,
-                )?))
-            }
-            // Strict profile (or non-checkpoint error variant): propagate the
-            // error so the runner sees the same failure mode as
-            // `checkpoint_and_exit_if_cancelled`. Returning `Ok(LoopExit::failed)`
-            // would silently mask `HostUnavailable` and break the strict
-            // require-final-checkpoint contract.
-            Err(error) => Err(error),
-        }
-    }
-
-    pub(super) async fn drain_user_inputs(
-        &self,
-        host: &(dyn AgentLoopDriverHost + Send + Sync),
-        mut state: LoopExecutionState,
-    ) -> Result<DrainedInputs, AgentLoopExecutorError> {
-        let batch = host
-            .poll_inputs(state.input_cursor.clone(), MAX_INPUT_DRAIN)
-            .await
-            .map_err(|_| AgentLoopExecutorError::HostUnavailable {
-                stage: HostStage::Input,
-            })?;
-        let (drained, ack_tokens, cancelled_reason_kind) =
-            consume_drainable_inputs(&batch, UserFacingInputDrainMode::Steering, &mut state)?;
-        Ok(DrainedInputs {
-            state,
-            drained,
-            ack_tokens,
-            cancelled_reason_kind,
-        })
-    }
-
-    pub(super) async fn drain_followup(
-        &self,
-        host: &(dyn AgentLoopDriverHost + Send + Sync),
-        mut state: LoopExecutionState,
-    ) -> Result<DrainedInputs, AgentLoopExecutorError> {
-        let batch = host
-            .poll_inputs(state.input_cursor.clone(), MAX_INPUT_DRAIN)
-            .await
-            .map_err(|_| AgentLoopExecutorError::HostUnavailable {
-                stage: HostStage::Input,
-            })?;
-        let (drained, ack_tokens, cancelled_reason_kind) =
-            consume_drainable_inputs(&batch, UserFacingInputDrainMode::FollowUp, &mut state)?;
-        Ok(DrainedInputs {
-            state,
-            drained,
-            ack_tokens,
-            cancelled_reason_kind,
-        })
     }
 }
