+++ crates/ironclaw_reborn/src/loop_driver_host/port_adapters.rs
@@ -1,4 +1,20 @@
-use super::*;
+use std::sync::Arc;
+
+use async_trait::async_trait;
+use ironclaw_turns::{
+    CheckpointStateStore, GetCheckpointStateRequest, GetLoopCheckpointRequest,
+    LoopCheckpointStateRef, LoopCheckpointStore, PutCheckpointStateRequest,
+    PutLoopCheckpointRequest, TurnCheckpointId,
+    run_profile::{
+        AgentLoopHostError, AgentLoopHostErrorKind, LoadCheckpointPayloadRequest,
+        LoadedCheckpointPayload, LoopCheckpointPort, LoopCheckpointRequest,
+        LoopHostMilestoneEmitter, LoopHostMilestoneSink, LoopInputAckToken, LoopInputBatch,
+        LoopInputCursor, LoopInputPort, LoopProgressEvent, LoopProgressPort, LoopRunContext,
+        LoopRunInfoPort, StageCheckpointPayloadRequest,
+    },
+};
+
+use super::turn_error_to_host_error;
 
 #[derive(Clone)]
 pub(super) struct NoExtraLoopInputPort {
